Fixed typos and made some cleanups in the async stuff

+++ src/control/control_hw.c
@@ -90,11 +90,11 @@ static int snd_ctl_hw_async(snd_ctl_t *ctl, int sig, pid_t pid)
 	}
 	if (sig < 0)
 		return 0;
-	if (fcntl(fd, F_SETSIG, sig) < 0) {
+	if (fcntl(fd, F_SETSIG, (long)sig) < 0) {
 		SYSERR("F_SETSIG failed");
 		return -errno;
 	}
-	if (fcntl(fd, F_SETOWN, pid) < 0) {
+	if (fcntl(fd, F_SETOWN, (long)pid) < 0) {
 		SYSERR("F_SETOWN failed");
 		return -errno;
 	}
